Incorporate MUI breadcrumbs into file. Rename Breadcrumbs to BreadcrumbsNav. Extract CSS to its own file.
